move css styles of support-tab to its own file
