+++ src/app/tabs/support-tab.js
@@ -1,60 +1,6 @@
 var yo = require('yo-yo')
 
-// -------------- styling ----------------------
-var csjs = require('csjs-inject')
-var styleGuide = require('remix-lib').ui.themeChooser
-var styles = styleGuide.chooser()
-
-var css = csjs`
-  .supportTabView {
-    height: 100vh;
-    padding: 2%;
-    padding-bottom: 3em;
-    display: flex;
-    flex-direction: column;
-    overflow: hidden;
-  }
-  .chat {
-    ${styles.rightPanel.supportTab.box_IframeContainer}
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    height: 85%;
-    padding: 0;
-  }
-  .chatTitle {
-    height: 40px;
-    width: 90%;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    margin-top: 15px;
-  }
-  .chatTitle:hover {
-    cursor: pointer;
-  }
-  .icon {
-    height: 70%;
-    margin-right: 2%;
-  }
-  .chatTitleText {
-    font-size: 17px;
-    font-weight: bold;
-  }
-  .chatTitleText {
-    opacity: 0.8;
-  }
-  .chatIframe {
-    width: 100%;
-    height: 100%;
-    transform: scale(0.9);
-    padding: 0;
-    border: none;
-  }
-  .infoBox {
-    ${styles.rightPanel.supportTab.box_SupportInfo}
-  }
-`
+var css = require('./styles/support-tab-styles')
 
 var infoText = yo`
   <div>
@@ -64,8 +10,6 @@ var infoText = yo`
     <a target="_blank" href='https://solidity.readthedocs.io/en/latest/'> Solidity</a>.
   </div>
 `
-
-module.exports = supportTab
 
 function supportTab (container, appAPI, events, opts) {
   var el = yo`
@@ -90,3 +34,5 @@ function supportTab (container, appAPI, events, opts) {
 function openLink () {
   window.open('https://gitter.im/ethereum/remix')
 }
+
+module.exports = supportTab
